Move the server and validation components to an .Internal namespace
