+++ src/OpenIddict.Server/Internal/OpenIddictServerEventService.cs
@@ -1,4 +1,10 @@
-﻿using System;
+﻿/*
+ * Licensed under the Apache License, Version 2.0 (http://www.apache.org/licenses/LICENSE-2.0)
+ * See https://github.com/openiddict/openiddict-core for more information concerning
+ * the license and the contributors participating to this project.
+ */
+
+using System;
 using System.Threading;
 using System.Threading.Tasks;
 using AspNet.Security.OpenIdConnect.Primitives;
@@ -6,12 +12,14 @@ using JetBrains.Annotations;
 using Microsoft.Extensions.DependencyInjection;
 using static OpenIddict.Server.OpenIddictServerEvents;
 
-namespace OpenIddict.Server
+namespace OpenIddict.Server.Internal
 {
     /// <summary>
-    /// Dispatches notifications by invoking the corresponding handlers.
+    /// Dispatches events by invoking the corresponding notification handlers.
+    /// Note: this API supports the OpenIddict infrastructure and is not intended to be used
+    /// directly from your code. This API may change or be removed in future minor releases.
     /// </summary>
-    public class OpenIddictServerEventService : IOpenIddictServerEventService
+    public class OpenIddictServerEventService
     {
         private readonly IServiceProvider _provider;
