+++ src/OpenIddict.Server/Internal/OpenIddictServerProvider.Serialization.cs
@@ -10,9 +10,14 @@ using AspNet.Security.OpenIdConnect.Primitives;
 using AspNet.Security.OpenIdConnect.Server;
 using JetBrains.Annotations;
 
-namespace OpenIddict.Server
+namespace OpenIddict.Server.Internal
 {
-    public partial class OpenIddictServerProvider : OpenIdConnectServerProvider
+    /// <summary>
+    /// Provides the logic necessary to extract, validate and handle OpenID Connect requests.
+    /// Note: this API supports the OpenIddict infrastructure and is not intended to be used
+    /// directly from your code. This API may change or be removed in future minor releases.
+    /// </summary>
+    public sealed partial class OpenIddictServerProvider : OpenIdConnectServerProvider
     {
         public override async Task DeserializeAccessToken([NotNull] DeserializeAccessTokenContext context)
         {
