+++ src/OpenIddict.Server/OpenIddictServerEvents.cs
@@ -1,4 +1,10 @@
-﻿using AspNet.Security.OpenIdConnect.Server;
+﻿/*
+ * Licensed under the Apache License, Version 2.0 (http://www.apache.org/licenses/LICENSE-2.0)
+ * See https://github.com/openiddict/openiddict-core for more information concerning
+ * the license and the contributors participating to this project.
+ */
+
+using AspNet.Security.OpenIdConnect.Server;
 using JetBrains.Annotations;
 
 namespace OpenIddict.Server
