+++ src/OpenIddict.Validation/IOpenIddictValidationEvent.cs
@@ -1,4 +1,10 @@
-﻿namespace OpenIddict.Validation
+﻿/*
+ * Licensed under the Apache License, Version 2.0 (http://www.apache.org/licenses/LICENSE-2.0)
+ * See https://github.com/openiddict/openiddict-core for more information concerning
+ * the license and the contributors participating to this project.
+ */
+
+namespace OpenIddict.Validation
 {
     /// <summary>
     /// Represents an OpenIddict validation event.
