+++ src/OpenIddict.Validation/Internal/OpenIddictValidationProvider.cs
@@ -5,7 +5,6 @@
  */
 
 using System;
-using System.ComponentModel;
 using System.Text;
 using System.Threading.Tasks;
 using AspNet.Security.OAuth.Validation;
@@ -13,17 +12,18 @@ using JetBrains.Annotations;
 using Microsoft.Extensions.DependencyInjection;
 using OpenIddict.Abstractions;
 
-namespace OpenIddict.Validation
+namespace OpenIddict.Validation.Internal
 {
     /// <summary>
-    /// Provides the logic necessary to extract, validate and handle OAuth2 requests.
+    /// Provides the logic necessary to extract and validate tokens from HTTP requests.
+    /// Note: this API supports the OpenIddict infrastructure and is not intended to be used
+    /// directly from your code. This API may change or be removed in future minor releases.
     /// </summary>
-    [EditorBrowsable(EditorBrowsableState.Never)]
-    public class OpenIddictValidationProvider : OAuthValidationEvents
+    public sealed class OpenIddictValidationProvider : OAuthValidationEvents
     {
-        private readonly IOpenIddictValidationEventService _eventService;
+        private readonly OpenIddictValidationEventService _eventService;
 
-        public OpenIddictValidationProvider([NotNull] IOpenIddictValidationEventService eventService)
+        public OpenIddictValidationProvider([NotNull] OpenIddictValidationEventService eventService)
             => _eventService = eventService;
 
         public override Task ApplyChallenge([NotNull] ApplyChallengeContext context)
@@ -44,7 +44,8 @@ namespace OpenIddict.Validation
                 {
                     throw new InvalidOperationException(new StringBuilder()
                         .AppendLine("The core services must be registered when enabling reference tokens support.")
-                        .Append("To register the OpenIddict core services, use 'services.AddOpenIddict().AddCore()'.")
+                        .Append("To register the OpenIddict core services, reference the 'OpenIddict.Core' package ")
+                        .Append("and call 'services.AddOpenIddict().AddCore()' from 'ConfigureServices'.")
                         .ToString());
                 }
 
@@ -117,7 +118,8 @@ namespace OpenIddict.Validation
                 {
                     throw new InvalidOperationException(new StringBuilder()
                         .AppendLine("The core services must be registered when enabling authorization validation.")
-                        .Append("To register the OpenIddict core services, use 'services.AddOpenIddict().AddCore()'.")
+                        .Append("To register the OpenIddict core services, reference the 'OpenIddict.Core' package ")
+                        .Append("and call 'services.AddOpenIddict().AddCore()' from 'ConfigureServices'.")
                         .ToString());
                 }
