+++ src/OpenIddict.Validation/OpenIddictValidationOptions.cs
@@ -5,6 +5,7 @@
  */
 
 using AspNet.Security.OAuth.Validation;
+using OpenIddict.Validation.Internal;
 
 namespace OpenIddict.Validation
 {
