+++ test/OpenIddict.Server.Tests/Internal/OpenIddictServerProviderTests.Session.cs
@@ -17,7 +17,7 @@ using Moq;
 using OpenIddict.Abstractions;
 using Xunit;
 
-namespace OpenIddict.Server.Tests
+namespace OpenIddict.Server.Internal.Tests
 {
     public partial class OpenIddictServerProviderTests
     {
