+++ test/OpenIddict.Server.Tests/Internal/OpenIddictServerProviderTests.cs
@@ -33,7 +33,7 @@ using OpenIddict.Abstractions;
 using OpenIddict.Core;
 using Xunit;
 
-namespace OpenIddict.Server.Tests
+namespace OpenIddict.Server.Internal.Tests
 {
     public partial class OpenIddictServerProviderTests
     {
