+++ test/OpenIddict.Validation.Tests/Internal/OpenIddictValidationInitializerTests.cs
@@ -16,7 +16,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Xunit;
 
-namespace OpenIddict.Validation.Tests
+namespace OpenIddict.Validation.Internal.Tests
 {
     public class OpenIddictValidationInitializerTests
     {
